feat: new index page to redirect on client side
